+++ lib/src/exceptions/couchdb_exception.dart
@@ -1,10 +1,15 @@
 import '../entities/db_response.dart';
 
+/// Exception that triggers when database respond with code 300 and above
 class CouchDbException implements Exception {
 
+  /// Returns instance with given error [code] and optional [response]
   CouchDbException(this.code, { this.response });
 
+  /// Status code
   int code;
+
+  /// Error response
   DbResponse response;
 
   @override