+++ lib/src/models/base/base_model.dart
@@ -1,7 +1,11 @@
 import '../../clients/base/couchdb_base_client.dart';
 
+/// Base model that hold connected [client] 
 abstract class BaseModel {
-  CouchDbBaseClient client;
-  
+
+  /// Creates instanse of [BaseModel] with given [client]
   BaseModel(this.client);
+
+  /// Instance of connected client
+  CouchDbBaseClient client;
 }