+++ lib/src/models/base/design_document_base_model.dart
@@ -3,6 +3,7 @@ import 'package:meta/meta.dart';
 import '../../clients/base/couchdb_base_client.dart';
 import 'base_model.dart';
 
+/// Class is under heavy development - don't use it for implementing your own methods!
 abstract class DesignDocumentBaseModel extends BaseModel {
 
   DesignDocumentBaseModel(CouchDbBaseClient client): super(client);