+++ lib/src/utils/includer_path.dart
@@ -1,1 +1,2 @@
+/// Method for including only non-null parameter to path
 String includeNonNullParam(String name, Object value) => value != null ? '$name=$value' : '';