+++ CHANGELOG.md
@@ -1,3 +1,10 @@
 # 0.0.1
 
-- Initial version
+- Initial version.
+- Implement server client for interacting with CouchDB.
+- Implement database-level methods.
+
+In future:
+
+- Implementation of document-level methods.
+- Client that work in browser.